requirements.txt
opencv-python~=4.5.1.48
numpy~=1.19.2
matplotlib~=3.4.1
pandas~=1.2.4
torchvision~=0.9.1
pyyaml~=5.4.1
tqdm~=4.60.0
setuptools~=52.0.0
pytorch~=1.8.1